requirements.txt
nextcord==2.2.0
asyncio
sqlite3
numpy
datetime
Pillow
topggpy